Moved DtoConnectionServer to CORE Project.
